+++ src/Generic.IrcClient/Dcc/DccSendMessage.cs
@@ -3,7 +3,7 @@
     public class DccSendMessage
     {
         public string FileName { get; set; }
-        public uint FileSize { get; set; }
+        public long FileSize { get; set; }
         public string IpAddress { get; set; }
         public int Port { get; set; }
     }
